requirements.txt
Flask
flask-cors
Pillow
rembg
numpy
zipfile36
shutil
uuid